Extract the post loader to a new class
to prepare for a merge
